requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.2
requests==2.31.0
beautifulsoup4==4.12.2
pandas==2.1.3
numpy==1.26.2
python-dotenv==1.0.0
google-generativeai==0.3.1
scikit-learn==1.3.2 